requirements.txt
streamlit
pandas
matplotlib
wordcloud
scikit-learn
spacy
keybert
requests
python-dotenv
beautifulsoup4